Include ZXing Android into your project.
ZXing is one of the most popular barcode scanning applications on the
market. They make it very easy for you to integrate into your
application via an intent but this means that your users must manually
install the application from the market.

Fortunately, the app is also open source so I will show you how to
cleanly build this capability into your project.